feat: add postres database to backend and make query and mutation work
